ETag response http status code is now 204 No Content. Fixed localtime handling.
The frontend sends ETag requests to see if the product has changed. The backend
will send the ETag response and now also sets the response code to 204 so that
logs are more easy to analyze. Also fixed the ETag not to change too easily
simply based on the wall clock.

Fixed localtime to be set to the proper location if the desired timezone is 'localtime'.
Boost.DateTime did not complain on the unknown zone, and times were handled incorrectly.