Fixed fmisid extraction from timeseries to ignore gaps in actual observations
Once the code was fixed to handle more than 1000 stations simultaneously, a previous
already failing test started failing more seriously. Turned out that when a bbox
is used to extract stations from an area, the code makes sure the same timesteps
exist for all stations. However, the code does not make sure the fmisid is still
defined for all the rows added due to fill the gaps.

Modified the code to extract the fmisid only once for the timeseries for a station,
the old code did it once for each row.